chore: moving SignerProxy from commons-ui to signer-protocol
Refs: XRDDEV-2412
